Enforce D7 core/UI boundary; relocate field-tracking hook out of math core (roadmap 0.11)
- Add ESLint no-restricted-imports rules scoping src/helpers/** and
  src/models/** to a framework-free core: no react, react-dom,
  react-number-format, @mui/*, @emotion/*, or *.tsx (UI component) imports.
  Makes the de-facto purity of the math layer a build failure, not a
  convention (decision D7).
- Move useFieldTracking (a React hook, the lone framework-dependent file
  under src/helpers) to src/hooks/, updating its two importers. This keeps
  the core import-clean and out of the Charter's 100% coverage scope.
