+++ src/loan/add-edit-loan.tsx
@@ -24,7 +24,7 @@ import {
   validateLoan,
 } from '../helpers/validation-helpers';
 import { fieldHelperText } from '../components/field-helper-text';
-import { useFieldTracking } from '../helpers/use-field-tracking';
+import { useFieldTracking } from '../hooks/use-field-tracking';
 
 export const AddEditLoan = (props: AddEditLoanProps) => {
   const [newLoan, setNewLoan] = useState<Loan>(emptyLoan);
